Add files via upload
Palm icon is proposed, as the tag:leaf_type supports the palm type. Several species that make up the urban trees are represented with this type of leaves. Although they are not very common trees in cities, the tag:leaf_type=broadleaved icon is not very representative for them.
The icon should be improved, I do not manage in the svg format.